Adds has_config Gem
This commit adds the has_config gem. It uses the gem to make the stats
column a config column. Since it is jsonb, the gem will now give us
access to a getter and setter for each key in the hash.

Also adds test to pokemon model to make sure the gem is functioning
correctly.